Crate e_cli

E-Cli

The command line tool for downloading E-Posts from a certain site.

It aims to be fast, and have verbose downloading console print outs.

What it can do:

Downloading Favourites of a user.

Downloading a single post from the site.

Downloading of multiple posts from a generated json file via get-pages

Downloading multiple pages

Usage

Downloading Favourites of a user with specific amount of posts.

e-tools.exe download-favourites {Username} -c {Number of Posts}

Downloading Random Favourites of a user

e-tools.exe download-favourites {Username} -r

Downloading a single post with lower quality

Note, this argument can be used with any of the commands the CLI has.

e-tools.exe --lower-quality download-post {Post Id}

Downloading pages of posts using the get-pages subcommand

First we get the pages via the subcommand:

e-tools.exe get-pages {Tags} -c {Number of Pages}

Then the command is as simple as it can be.

e-tools.exe download-posts posts.json

posts.json was generated in the root dir when get-pages was ran

Modules

cli

The cli definitions using [clap] derive

commands

The module with all the command functions.

funcs

Special functions for the cli that would otherwise be reused multiple times.

type_defs

All Type Definitions the program uses.

Functions

main 🔒

Main function, handles the cli arguments.